v0.0.3

+++ src/main/java/be/kuleuven/cs/jli40d/client/Client.java
@@ -65,7 +65,6 @@ public class Client extends JFrame implements ActionListener
         setDefaultCloseOperation( JFrame.EXIT_ON_CLOSE );
         setResizable( false );
 
-        /*
         loginPanel = new JPanel();
 
         usernameField = new JTextField( "Username" );
@@ -84,7 +83,6 @@ public class Client extends JFrame implements ActionListener
         loginPanel.add( login );
 
         add( loginPanel );
-        */
         setVisible( true );
     }
 
@@ -173,8 +171,8 @@ public class Client extends JFrame implements ActionListener
                     String value2   = nrOfPlayersField.getText();
                     try
                     {
-                        int nrOfPlayers = Integer.valueOf( value2 );
-                        int id          = lobbyHandler.makeGame( token, gameName, nrOfPlayers );
+                        int  nrOfPlayers = Integer.valueOf( value2 );
+                        long id          = lobbyHandler.makeGame( token, gameName, nrOfPlayers );
                         updateLobbyList();
                     }
                     catch ( RemoteException e1 )
@@ -240,7 +238,7 @@ public class Client extends JFrame implements ActionListener
 
     }
 
-    private void joinGame( int id )
+    private void joinGame( long id )
     {
         try
         {
