+++ src/main/java/be/kuleuven/cs/jli40d/core/LobbyHandler.java
@@ -41,7 +41,7 @@ public interface LobbyHandler extends Remote, Serializable
      * @throws UnableToCreateGameException When the game cannot be created for some reason (like exceeded limits).
      * @throws RemoteException
      */
-    int makeGame( String token, String gameName, int numberOfPlayers ) throws RemoteException, InvalidTokenException, UnableToCreateGameException;
+    long makeGame( String token, String gameName, int numberOfPlayers ) throws RemoteException, InvalidTokenException, UnableToCreateGameException;
 
     /**
      * Join a game with an id, either provided by the {@link #makeGame} or {@link #currentGames} method.
@@ -59,5 +59,5 @@ public interface LobbyHandler extends Remote, Serializable
      * @throws InvalidTokenException     When the token is invalid (expired or not found).
      * @throws RemoteException
      */
-    Game joinGame( String token, int gameId ) throws RemoteException, UnableToJoinGameException, InvalidTokenException;
+    Game joinGame( String token, long gameId ) throws RemoteException, UnableToJoinGameException, InvalidTokenException;
 }
