+++ src/main/java/be/kuleuven/cs/jli40d/server/application/GameListHandler.java
@@ -12,11 +12,11 @@ import java.util.List;
  */
 public interface GameListHandler extends Serializable
 {
-    void add(Game game);
+    void add( Game game );
 
     int nextID();
 
-    Game getGameByID( int id ) throws GameNotFoundException;
+    Game getGameByID( long id ) throws GameNotFoundException;
 
     List<Game> getAllGames();
 }
